Update receipts.sql
fee scalar to decimal

+++ schemas/aws/receipts.sql
@@ -12,7 +12,7 @@ CREATE EXTERNAL TABLE IF NOT EXISTS receipts (
     l1_fee BIGINT,
     l1_gas_used BIGINT,
     l1_gas_price BIGINT,
-    l1_fee_scalar BIGINT
+    l1_fee_scalar DECIMAL
 )
 PARTITIONED BY (block_date STRING)
 ROW FORMAT SERDE 'org.apache.hive.hcatalog.data.JsonSerDe'
